Reworked the entire site so that it uses a new new, more detailed JSON documentation structure that includes more precise information about models as well as data types for parameters & fields.
Added PHP and cURL code generators to output code based on the current
state of the API builder.

Started using Typescript to make it easier to code against the much more detailed definitions file that drives the site.

Added a model browser that details information about all the models and their fields.

Updated the method view page to include more information about allowed values and data types that each parameter accepts.  Added static documentation explaining how the various parameter types are structured and examples for using each.  For parameters that are bound to a model or field, documentation for those relations is now linked.

Reworked the API builder to take detailed data types into consideration, ensuring that the HTML inputs are able to collect the appropriate values.  The data object and filter parameters types were most heavily modified to guide user inputs.

Removed the method list view as it was redundant.

Removed unused test code framework and libraries (clearing them out until we find time to make use of them), reducing this repository to a static set of javascript / HTML.